# src/Output/openapi.yaml
openapi: 3.0.0
info:
  title: 'Calendar App'
  description: 'This API is designed to facilitate the management of a comprehensive calendar application, encompassing a broad range of functionalities including user management, event scheduling, course allocation, group coordination, department organization, and timetable structuring. It aims to provide an integrated platform for educational institutions to efficiently organize and access academic and administrative information.'
  version: 1.0.3
servers:
  -
    url: 'http://localhost/web_project/src/'
paths:
...
  '/events/{id}':
    get:
      tags:
        - Events
      summary: 'Retrieve an event by ID'
      description: 'Retrieves an event by its ID.'
      operationId: 1fbd657c92defab0d86e26ca3d313cec
      parameters:
        -
          name: id
          in: path
          description: 'ID of the event to retrieve'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 'Successful operation'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '404':
          description: 'Event not found'
          content:
            application/json:
              schema:
                properties:
                  status: { type: string, example: error }
                  message: { type: string, example: 'Event not found' }
                type: object
    put:
      tags:
        - Events
      summary: 'Update an event by ID'
      description: 'Updates an event by its ID.'
      operationId: 6c636e315a7f5907c3f51cebe18f9021
      parameters:
        -
          name: id
          in: path
          description: 'ID of the event to update'
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: 'Updated event data'
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Event'
      responses:
        '200':
          description: 'Successful operation'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '404':
          description: 'Event not found'
          content:
            application/json:
              schema:
                properties:
                  status: { type: string, example: error }
                  message: { type: string, example: 'Event not found' }
                type: object
    delete:
      tags:
        - Events
      summary: 'Delete an event by ID'
      description: 'Deletes an event by its ID.'
      operationId: c6c527508ad10c073bb2a75e2ac5d588
      parameters:
        -
          name: id
          in: path
          description: 'ID of the event to delete'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: 'Successful operation'
          content:
            application/json:
              schema:
                properties:
                  status: { type: string, example: success }
                  message: { type: string, example: 'Event deleted' }
                type: object
        '404':
          description: 'Event not found'
          content:
            application/json:
              schema:
                properties:
                  status: { type: string, example: error }
                  message: { type: string, example: 'Event not found' }
                type: object
...
  /matters:
    get:
      tags:
        - Matters
      summary: 'List all matters or filter by matter ID'
      operationId: 5b7ceecb6d703e997c459d650953f01e
      parameters:
        -
          name: id
          in: query
          description: 'ID of the matter to retrieve'
          required: false
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 'A list of matters'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Matter'
        '404':
          description: 'A matter with the specified criteria was not found'
    post:
      tags:
        - Matters
      summary: 'Create a new matter'
      description: 'Creates a new matter.'
      operationId: 38eb8ba0ac98a75d21d81e3dc63127ac
      requestBody:
        description: 'Matter data to create'
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Matter'
      responses:
        '201':
          description: 'Successful operation'
          content:
            application/json:
              schema:
                properties:
                  status: { type: string, example: success }
                  data: { $ref: '#/components/schemas/Matter' }
                type: object
        '400':
          description: 'Bad request'
          content:
            application/json:
              schema:
                properties:
                  status: { type: string, example: error }
                  message: { type: string, example: 'Bad request' }
                type: object
  '/matters/{id}':
    get:
      tags:
        - Matters
      summary: 'Retrieve a matter by ID'
      description: 'Retrieves a matter by its ID.'
      operationId: 4b2068ec501970b354ed2e76bc9fc590
      parameters:
        -
          name: id
          in: path
          description: 'ID of the matter to retrieve'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 'Successful operation'
          content:
            application/json:
              schema:
                properties:
                  status: { type: string, example: success }
                  data: { $ref: '#/components/schemas/Matter' }
                type: object
        '404':
          description: 'Matter not found'
          content:
            application/json:
              schema:
                properties:
                  status: { type: string, example: error }
                  message: { type: string, example: 'Matter not found' }
                type: object
    put:
      tags:
        - Matters
      summary: 'Update a matter by ID'
      description: 'Updates a matter by its ID.'
      operationId: 8ba175f9346e8cbb5f29b4a006820277
      parameters:
        -
          name: id
          in: path
          description: 'ID of the matter to update'
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: 'Matter data to update'
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Matter'
      responses:
        '200':
          description: 'Successful operation'
          content:
            application/json:
              schema:
                properties:
                  status: { type: string, example: success }
                  data: { $ref: '#/components/schemas/Matter' }
                type: object
        '404':
          description: 'Matter not found'
          content:
            application/json:
              schema:
                properties:
                  status: { type: string, example: error }
                  message: { type: string, example: 'Matter not found' }
                type: object
    delete:
      tags:
        - Matters
      summary: 'Delete a matter by ID'
      description: 'Deletes a matter by its ID.'
      operationId: 229280eff53786fe29c33ae5889619b9
      parameters:
        -
          name: id
          in: path
          description: 'ID of the matter to delete'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: 'Successful operation'
          content:
            application/json:
              schema:
                properties:
                  status: { type: string, example: success }
                  message: { type: string, example: 'Matter deleted' }
                type: object
        '404':
          description: 'Matter not found'
          content:
            application/json:
              schema:
                properties:
                  status: { type: string, example: error }
                  message: { type: string, example: 'Matter not found' }
                type: object
...
components:
  schemas:
    Event:
      title: Event
      description: 'Event entity schema'
      required:
        - id
        - matterId
        - teacherId
        - groupId
        - roomId
        - startTime
        - endTime
        - type
      properties:
        id:
          description: 'Unique identifier for the event'
          type: integer
          format: int64
        matterId:
          description: 'Identifier for the matter associated with the event'
          type: integer
        teacherId:
          description: 'Identifier for the teacher associated with the event'
          type: integer
        groupId:
          description: 'Identifier for the group associated with the event'
          type: integer
        roomId:
          description: 'Identifier for the room associated with the event'
          type: integer
        startTime:
          description: 'Start time of the event (in ISO 8601 format)'
          type: string
          format: date-time
        endTime:
          description: 'End time of the event (in ISO 8601 format)'
          type: string
          format: date-time
        type:
          description: 'Type of the event'
          type: string
          enum:
            - course
            - seminary
            - laboratory
      type: object
...
    Matter:
      title: Matter
      description: 'Matter entity schema'
      required:
        - id
        - title
        - type
      properties:
        id:
          description: 'Unique identifier for the matter'
          type: integer
          format: int64
        title:
          description: 'Title of the matter'
          type: string
        type:
          description: 'Type of the matter'
          type: string
          enum:
            - mandatory
            - optional
            - elective
      type: object
...